🎨 route 관련 코드 정리
