+++ src/page/index.js
@@ -2,3 +2,4 @@ export { default as HomePage } from "./HomePage";
 export { default as LoginPage } from "./LoginPage";
 export { default as NotFoundPage } from "./NotFoundPage";
 export { default as ProfilePage } from "./ProfilePage";
+export { default as ErrorBoundaryPage } from "./ErrorBoundaryPage";
